src/Document/requirements.txt
pika==1.3.0
PyPDF2==3.0.1
psycopg2-binary==2.9.6
nltk
sentence-transformers